// lib/types.ts
import { ProductImage } from "@prisma/client";
export type ProductSpecRow = { col1: string; col2: string };

// lib/types.ts
export interface Category {
  id: string;
  name: string;
  name_zh: string | null;
  slug: string;
  createdAt?: Date;
  updatedAt?: Date;
}

export interface Product {
  id: string;
  name: string;
  name_zh: string | null;
  description: string | null;
  specs_zh?: unknown | null;

  featured: boolean;
  slider: boolean;
  categoryId: string;
  category: Category;
  images: ProductImage[];
  createdAt: Date;
  updatedAt: Date;
}

export interface CartItem {
  id: string;
  productId: string;
  name: string;

  image: string;
  quantity: number;
}